+++ base/src/iosMain/kotlin/utils/DataTypeUtils.kt
@@ -25,7 +25,6 @@ import kotlinx.cinterop.memScoped
 import kotlinx.cinterop.reinterpret
 import platform.Foundation.NSData
 import platform.Foundation.create
-import platform.darwin.NSUInteger
 
 /**
  * Converts a [NSData] to its corresponding [ByteArray]
@@ -42,7 +41,7 @@ fun NSData.toByteArray(): ByteArray {
  * @return Th [NSData] corresponding to this [ByteArray]
  */
 fun ByteArray.toNSData(): NSData = memScoped {
-    return NSData.create(bytes = allocArrayOf(this@toNSData), length = this@toNSData.size.toULong() as NSUInteger)
+    return NSData.create(bytes = allocArrayOf(this@toNSData), length = this@toNSData.size.toULong())
 }
 
 /**
